Insert status as QUINT64 in mailmessages database, not int
SQlite is storing signed 64bit integers. With the static_cast<int>, the value
is truncated. It is not necessary to cast the value as a qint64 neither since the
sign-bit is not used anyway to store a flag (there are maximum 63 flags).

+++ src/libraries/qmfclient/qmailstore_p.cpp
@@ -5989,7 +5989,7 @@ QMailStorePrivate::AttemptResult QMailStorePrivate::attemptAddMessage(QMailMessa
     values.insert(QLatin1String("recipients"), recipients.join(QLatin1String(",")));
     values.insert(QLatin1String("subject"), metaData->subject());
     values.insert(QLatin1String("stamp"), QMailTimeStamp(metaData->date()).toUTC());
-    values.insert(QLatin1String("status"), static_cast<int>(metaData->status()));
+    values.insert(QLatin1String("status"), metaData->status());
     values.insert(QLatin1String("parentaccountid"), metaData->parentAccountId().toULongLong());
     values.insert(QLatin1String("mailfile"), ::contentUri(*metaData));
     values.insert(QLatin1String("serveruid"), metaData->serverUid());
